Generate config file only if it don't exists. (msi)

+++ wix/replace/replace_windows.go
@@ -21,8 +21,12 @@ func main() {
 		log.Fatal(err)
 	}
 
-	err = ioutil.WriteFile(outFile, []byte(strings.Replace(string(content), oldStr, newStr, -1)), 0644)
-	if err != nil {
-		log.Fatal(err)
+	_, err = os.Stat(outFile)
+	outFileIsExists := err == nil
+	if !(outFileIsExists) {
+		err = ioutil.WriteFile(outFile, []byte(strings.Replace(string(content), oldStr, newStr, -1)), 0644)
+		if err != nil {
+			log.Fatal(err)
+		}
 	}
 }
